refactor(server): expose component sources edge-safely
